Guild Gifts auto collector
This macro is designed to work with Lords Mobile. If you haven't reached VIP level 12, you are probably developing carpal tunnel syndrome from collecting all the guild gifts, especially if you are in a big guild that kills a lot of monsters. This macro will make your life easier. Instead of hundreds of clicks, just a few and all your guild gifts will be collected automatically.

Notes:
Make sure you are on the Guild Gifts view (the view with all the gifts) when you run the app.
May need to run more than once if you have hundreds to collect.