Shelter 12 - Auto shelter 12 hour
This FRep macro for Lords Mobile will auto shelter your leader and troops. It will reset a 12 hour shelter every 3600 seconds. The time interval can be adjusted. 
IMPORTANT: The shelter icon MUST be over the GO button. The blue button telling you what building to upgrade next. The screen must remain in this position for the macro to work. Otherwise, it will not click the shelter icon. and strange behavior could occur.